Move KlapTransport into transports package
