+++ kasa/transports/klaptransport.py
@@ -57,12 +57,17 @@ from cryptography.hazmat.primitives import padding
 from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
 from yarl import URL
 
-from .credentials import Credentials
-from .deviceconfig import DeviceConfig
-from .exceptions import AuthenticationError, KasaException, _RetryableError
-from .httpclient import HttpClient
-from .json import loads as json_loads
-from .protocol import DEFAULT_CREDENTIALS, BaseTransport, get_default_credentials, md5
+from kasa.credentials import Credentials
+from kasa.deviceconfig import DeviceConfig
+from kasa.exceptions import AuthenticationError, KasaException, _RetryableError
+from kasa.httpclient import HttpClient
+from kasa.json import loads as json_loads
+from kasa.protocol import (
+    DEFAULT_CREDENTIALS,
+    BaseTransport,
+    get_default_credentials,
+    md5,
+)
 
 _LOGGER = logging.getLogger(__name__)
